app/requirements.txt
crispy-bootstrap5==2025.6
Django==6.0.2
django-crispy-forms==2.5
django-cors-headers==4.9.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.16.1
dj-rest-auth==7.1.1
drf-social-oauth2==3.2.0
drf-spectacular==0.29.0
requests==2.32.5
